requirements.txt
bs4
feedparser
flask
frozen-flask
gunicorn
python-slugify
pyyaml
